Refactor postgis Dockerfile to accept a PGVER argument
Right now you can pass from 9.0 to 9.4 as values,
but only 9.3 and 9.4 have been tested.
The build.sh script passes 9.3 to be backward compatible.

The Dockerfile was also optimized to reduce the intermediary
images while appending lines to PostgreSQL configuration and
a fix for forced-pull of PostgreSQL-9.4 was included, which
closes #13