Add a create_storage_containers.sh script under dev2.systemapic.com
The script takes the path to a docker-compose.yml, aiming to be
generic. It then guesses on the volume path to be exported based
on the name of the container referenced by docker-compose.yml.

Extraction of volume name is not robust as it relies on the
following assumptions:

 1: There is a single volume after "volumes-from:" line
 2: The single volume is right below the "volumes-from" line

Also add a suggestion in ./restart.sh in case containers are missing

See #31